fix
git-svn-id: http://bgslibrary.googlecode.com/svn/trunk@58 7e9cb7ec-d131-0c6a-8031-cf146dc18c44

+++ package_bgs/lbp_mrf/MEImage.cpp
@@ -1286,7 +1286,7 @@ int MEImage::NeighbourhoodCounter(int startx, int starty,
 void MEImage::GradientVector(bool smooth, int x, int y, int mask_size, int& result_x, int& result_y)
 {
   int Results[8];
-  int DiagonalMaskSize = (int)((float)mask_size / std::sqrt((double)2));
+  int DiagonalMaskSize = (int)((float)mask_size / sqrt(2));
 
   if (ME_CAST_TO_IPLIMAGE(cvImg)->nChannels > 1)
   {
